Revert "添加oss配置并将上传视频存放到oss"
