+++ models/app.ini
@@ -5,9 +5,7 @@ user     = mysql
 password = mysql
 database = douyin
 
-[oss]
-endpoint        = oss-cn-shenzhen.aliyuncs.com
-accessKeyId     = LTAI5tFSSHJVuM4K37KjpGND
-accessKeySecret = c8FWiwZoRBaZapDJD8vGaNC3dcYSnv
-bucket          = dousheng1
+; [redis]
+; ip   = 127.0.0.1
+; port = 6379
